+++ CHANGELOG.md
@@ -17,7 +17,9 @@
     - `Report`
     - `Scope`
     - `State`
-- `State::Framework#abort` -- Fixed exception message.
+- `State::Framework`
+    - Added `done?`
+    - `#abort` -- Fixed exception message.
 - Checks
     - Active
         - `sql_injection` -- Slight payload update to catch double-quote cases.